+++ pingouin/__init__.py
@@ -10,4 +10,4 @@ from .circular import *
 from .pairwise import *
 from .power import *
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"